+++ tools/onnx-graphsurgeon/onnx_graphsurgeon/ir/graph.py
@@ -478,6 +478,7 @@ class Graph(object):
         error_ok=True,
         flatten_subgraphs=True,
         size_threshold=None,
+        should_exclude_node=None,
     ):
         """
         Folds constants in-place in the graph. The graph must be topologically sorted prior to
@@ -525,11 +526,18 @@ class Graph(object):
                     the model size, it may be desirable to skip folding them and allow them to be computed
                     at runtime.
                     Defaults to None.
+            should_exclude_node (Callable[[gs.Node], bool]):
+                    A callable that accepts an onnx-graphsurgeon node from the graph and reports whether it should
+                    be excluded from folding. This is only called for nodes which are otherwise foldable.
+                    Note that preventing a node from being folded also prevents its consumers from being folded.
+                    Defaults to a callable that always returns False.
 
         Returns:
             self
         """
         from onnx_graphsurgeon.exporters.onnx_exporter import dtype_to_onnx, export_onnx
+
+        should_exclude_node = misc.default_value(should_exclude_node, lambda node: False)
 
         PARTITIONING_MODES = [None, "basic", "recursive"]
         if partitioning not in PARTITIONING_MODES:
@@ -671,7 +679,7 @@ class Graph(object):
                         foreign_tensors = attr._foreign_tensors().values()
                         all_subgraph_foreign_tensors_const &= all_tensors_const(foreign_tensors)
 
-                return all_subgraph_foreign_tensors_const
+                return all_subgraph_foreign_tensors_const and not should_exclude_node(node)
 
             # Walks along the outputs of graph_constants to see if they can also be computed statically.
             # Since the graph is topologically sorted, this should find all constant nodes in the graph.