+++ tools/onnx-graphsurgeon/CHANGELOG.md
@@ -3,6 +3,10 @@
 Dates are in YYYY-MM-DD format.
 
 ## v0.3.25 (2022-10-14)
+### Added
+- Added a `should_exclude_node` parameter to `fold_constants` to allow for excluding nodes
+    from constant folding.
+
 ### Fixed
 - Fixed a bug where `fold_constants` would fold quantization nodes, which are intended to be executed
     at runtime even though they are computable beforehand.